pyproject.toml
[project]
name = "mini-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.25",
    "langchain-groq>=1.1.2",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.5",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.3.3",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.43.0",
]
